Refactor test harness database initialization and schema management
